shelldriver: expect empty password login despite password is provided
When a system was configured to not expect a password (maybe for
debugging purpose) but the shelldriver configuration had a password set,
login failed as the shelldriver only expected a "Password: " line.

This is handled now by also matching for a prompt even when a password
is configured. Additionally, an explicit Exception will be raised now
when no password is configured but the system expects one.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>