ShellDriver: implement file transfer support via XMODEM
This implementation starts the respective sender or receiver on the
target and then uses tehmaze's xmodem Python implementation [1] to
transfer files over the serial console. It can work with busybox's rx
(only for receiving) or lrz/lsz from lrzsz [2] (for receiving and
sending) on the target, and autodetects which commands are available.

XMODEM is a very simple, but also very historic protocol with some
quirks. See the comment in labgrid.driver.ShellDriver._put() for some
insight and workarounds.

Unfortunately, after the peer on the target has been started, there is
no way to stop it if the transfer fails and get back to a shell prompt,
since the peer on the target interprets everything on the console as
XMODEM messages. lsz's --min-bps option does not seem to work very good
(but is included nevertheless for the sake of trying), so powercycling
the target is possibly the best way to get rid of such situations.

This patch also includes support for uploading and running scripts onto
the target to make it easier for the user to automate more complex
scenarios.

[1]: https://github.com/tehmaze/xmodem
[2]: https://ohse.de/uwe/software/lrzsz.html

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>