multiple: explicitly set attr.s cmp parameter
We mostly don't need the compare functionality of attrs. Disable it to fall back
to object id hashes (similar to python standard classes). The only exception is
the ResourceMatch class, since we collect these into a list and need to remove
them later on. The list remove function requires the implementation of the
compare functions, explicitly enable it for the ResourceMatch class.

Fixes #106

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>