remote/client: local configuration target and state support
The _get_target function now tries to retrieve a place from the local
configuration, for this to work remote and local target name have to match.

The client now supports the STATE environment variable and -s/--state command
line parameter. This will bring the target into the requested strategy state
before executing the command.

We also rework all commands which require a specific driver to check wether the
configuration already supplies this driver and use it instead.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>