remote: support named matches
Matches can now have an optional name. When a resource matches a named
ResourceMatch, it will be made available on the client under that name.
This is useful when using multiple resources of the same class in one
target.

Note that clients and coordinator must be updated together, as the
older clients do not understand the 'rename' field.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>