remote: replace autoping with explicit polling
This fixes autobahn timeouts during long-running tests.

The mail problem is that the asyncio eventloop of the remote client is
not running while executing test cases. That causes the autoping
requests from the crossbar to time out, which triggers a disconnect.

We can't easily use autoping only for the exporters, so this implements
explicit polling from the coordinator to each exporter. It uses a new
'version' procedure for that, which is also useful for debugging.

The long-term fix is probably to move the autobahn client event loop
into a python thread, but that is a much larger change with higher risk.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>