target: driver, resource and protocol retrieval by string
The class functions get_driver, get_active_driver and get_resource now accept
a string as a parameter as well and use an internal lookup table to convert from
strings to a bound class. Additionally the lookup table saves protocols as well,
allowing driver retrieval by protocol in getitem:

  shell = target["CommandProtocol"]

removing a lot of the imports necessary in test modules.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>