requirements: split up requirements files
The requirements.txt file is used for the new installation process,
crossbar-requirements is used to fix the crossbar dependencies as well.
To use labgrid as an application, we pin the requirements of labgrid to known
tested values.
We also remove crossbar from extra_require, since we have a new installation
process for the coordinator.
We also split out requirements files for every optional feature.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>