driver/shelldriver: rework _await_login
This commit changes the behavior of _await_login in two ways:

Enable Console:
===============

This feature can be used for embedded devices that do not - by default -
start a console on the serial terminal.  The 'console_ready'-regex is
used to identify when the kernel is ready to start a console. Afterwards
<enter> is send to the device. The shelldriver now waits for a prompt or login.

Tested with a 4.4-kernel on a LEDE-based device.

If 'console_ready' is not set _await_login() should behave like before.

Wait for silence before probing DUT state:
==========================================

This commit addresses https://github.com/labgrid-project/labgrid/issues/176
by removing the auto-sending of a newline when waiting for a prompt.

_await_login() still tries to match for a prompt, login or optional
'activate console'-message. If none of these is detected the driver will wait
for await_login_timeout seconds of silence on the serial before sending a
newline to the DUT.

Trying to activate a prompt will be continued for a maximum of login_time
seconds before raising an Exception.

Signed-off-by: Chris Fiege <chris@tinyhost.de>