driver: add SerialDigitalOutoutDriver
The SerialDigitalOutoutDriver sits on top of a SerialDriver and uses the
underlying pyserial object to control a flow control line of a serial
port as if it were general-purpose outputs.

With this driver a free serial port can be used to control
a general purpose output for eg. choosing of a boot-target.

Signed-off-by: Chris Fiege <chris@tinyhost.de>