target: implement priorities
Implement priority checking in case the target finds multiple drivers. If the
targets have a driver with a higher priority, choose that one, otherwise raise a
NoDriverFoundError with additional Information.

For drivers without a priority class variable we default to 0, when we reach the
Protocol during the iteration over the MRO.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>