labgrid/environment: do not drop error information by re-raising error
Config already raises NoConfigFound error with a proper message set.
There is no need to re-raise it.
Furthermore, it also raises InvalidConfigError which will be hidden to
the upper layers by catching and re-raising untargeted.

This is fixed here by simply removing the try-catch-clause around the
Config() constructor call.

Now, that we do not unconditionally catch all errors, former yaml config
errors that were re-reised as NoConfigFoundErrors will now be raised as
YAMLError as they are not catched in config.py.

Thus, to be consistent, we catch YAMLErrors an re-raise them as
InvalidConfigErrors with their original error message and expect this in
the corresponding test case.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>