remote: client: try already instantiated bootstrap drivers first
I want to use imx-usb-loader with a remote i.MX28 (MXS) device with the
following env.yaml:

  targets:
    main:
      resources:
        RemotePlace:
        name: my-remote-place
      drivers:
        NetworkPowerDriver: {}
        SerialDriver: {}
        IMXUSBDriver: {}

In this case the remote place only has a MXSUSBDriver because of the
USB ID matching in udev.py, so the client tries to instantiate a
MXSUSBDriver first, suppressing our attempt to override it with
IMXUSBDriver in our YAML file.

To get rid of this, teach the client to use the first available
bootstrap driver it can find, and only then try to instantiate the
others.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>