driver/networkusbstoragedriver: fix concatinating tuple to list in get_size()
The cat command was moved to its own variable during linting, but
accidentally to a tuple and not a list. This leads to an error while
concatinating with self.storage.command_prefix:

  TypeError: can only concatenate list (not "tuple") to list

Fix that.

Another occurrence of the same error was fixed by 2544d1d1
("driver/networkusbstoragedriver: fix concatinating tuple to list")

Fixes: 533cec4 ("linting: do miscellaneous linting")
Signed-off-by: Bastian Stender <bst@pengutronix.de>