remote/exporter: set LOCAL option for ser2net exported serial ports
Identify ports which are exported over ser2net as local ports, ignoring the
modem control lines.

We had a device which uses a full rs232 connection instead of the usual serial
to ttl usb converter. Somewhere in the boot process the device side closed the
port and signaled this via the modem control lines, reulting in an EOF in
ser2net.
Configure the ser2net subprocesses to assume that all serial ports are local
ports, telling the kernel to ignore the control lines.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>