pytestplugin: move env creation into pytest_configure call
To filter the tests in the pytest collect stage, we need to parse the
environment file and create the environment earlier in pytest.
Before it was created when the first test called the fixture.
We now always create the environment in the pytest configure stage and pass the
created environment to the fixture.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>