add docker container support
Implement labgrid additions to interact with a docker daemon using the
"Docker SDK for Python" library.

Driver Addition:
- DockerDriver to interact with the docker daemon.

Resource Additions:
- DockerDaemon to locate the docker daemon (through url) and to
  create a NetworkService out of the docker container created through
  interaction with the docker daemon.
- DockerManager to manage the DockerDaemon (a ManagedResource).

Strategy Addition:
- DockerStrategy to control start and stop of the containers.

Use of this features are illustrated by the demo in examples/docker.

The test in tests/test_docker.py is used by tox and checks behaviour
both while a docker daemon is present and when it is rather mocked
away.

Signed-off-by: Michael Due Jensen <mdj@triax.dk>