util/ssh: copy also symlink referents to remote
In case an image file is given as a symlink pointing to the actual image
the rsync command skips the file:

  skipping non-regular file "myimage.img"

Fix this by actually copying the symlink referent to the remote.

This is important for Yocto builds. The resulting images contain a
timestamp in the file name. For each image a symlink with a predictable
name is created. So using this symlink is the easiest way of
referencing a Yocto image in an automated build and test chain.

See: https://www.yoctoproject.org/docs/1.8/ref-manual/ref-manual.html#images-dev-environment

Signed-off-by: Bastian Krause <bst@pengutronix.de>