remote/client: add --sort-by-matched-place-change for resources
'labgrid-client resources --sort-by-matched-place-change' shows
resources along with places (and their last changed date) matching the
respective resource. The places per resource are sorted by changed date
beginning with the most recent one. So the first listed place per
resource can be interpreted as the "last changed date" of the resource.
The list of resources is sorted by exactly that "last changed date" of
the resource.

This enables the user to
  a) identify resources (e.g. power ports) not used for a long time
  b) identify resources matched by multiple places

Use case a) is especially useful when combined with a match given after
the "resources" subcommand to identify resources by
exporter/group/class/name or even combined.

'labgrid-client --verbose resources' shows resources per exporter and
group. As this obviously contradicts the use case of
--sort-by-matched-place-change there is currently no verbose version
for sorted output.

Signed-off-by: Bastian Krause <bst@pengutronix.de>