resource/udev: use udev in thread and timeout
The udev monitor is now run in a thread and drains the udev socket of
events. The events are put into a Queue, which is worked on for
approximately 100ms. This should fix the exporter not responding to
pings from the coordinator, which led to the exporter being kicked.
Instances where the udev socket are filled up should also be reduced,
since a thread is now used to drain the socket.

Signed-off-by: Rouven Czerwinski <rouven@czerwinskis.de>