driver/networkusbstoragedriver: support images configuration
Previously the NetworkUSBStorageDriver only took absolute paths in the
write_image function. Allow the specification of an image attribute which is
used as a fallback if no explicit filename is given.
This brings this driver in line with other drivers which can also retrieve their
files from the environment configuration.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>