util/managedfile: heuristically detect NFS to avoid copy
With #381, the fastboot client command would always attempt rsync'ing
the local file to the remote. It can be preferable to avoid this copy
if we the file is located on a NFS accessible to both the local and
remote system. This improves user experience because uploading to DUT
and download from labgrid-client system can happen in parallel.

To detect if the file is the same we use stat(1) on the same path on
both local and remote system and compare inode number, total size
and birth/modification timestamps.

This approach requires GNU coreutils' stat(1) to be available on
both local and remote system. Besides parsing ls -l, there seems
to be no fully POSIX alternative to get the required information
without writing C.

If detect_nfs is False, stat(1) is unavailable or incompatible with
--printf, or the stat(1) output of both systems don't match,
the rsync would be attempted as usual.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>