driver/shelldriver: handle devices that do not have a login prompt
For example when using a minimal initramfs system, we might not even
have a login prompt to catch. With the current implementation the only
way of handling this is to set 'login_prompt' to a string that will
never be machted which is quite inconvenient.

A better handling is to make the 'login_prompt' optional, respectively
allow passing 'login_prompt=""'.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>