driver: fastbootdriver: echo command before executing it on run()
The output of commands executed via 'fastboot oem exec <cmd>' cannot be
seen on the host. The command itself cannot be seen via serial on the
barebox shell.

This adds the possibility to echo the command before executing it. This
allows to see which commands lead to which (error) messages on the
serial interface without looking at labgrid's log files. This is
especially useful for scenarios without a SerialDriver, but with a
non-regular serial terminal connected for debugging only.

Signed-off-by: Bastian Krause <bst@pengutronix.de>