util/proxy: implement forced proxy mode
When the client cannot access the exporters and coordinator directly,
we can now use the ProxyManager to create SSH port forwards
automatically, even when the exporter is not configured to require
proxied connections.

Also add code to the client to set the forced proxy hostname. To always
use the same proxy host, use the LG_PROXY environment variable.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>