driver: fastbootdriver: don't disable sparsing by default
By default, the sparse threshold is computed by querying the
target's max-download-size variable.

With #416, labgrid adopted the behavior of passing -S 0 as
default value to fastboot. This is problematic as different
fastboot versions interpret -S 0 differently:

- Fastboot releases containing 542370dead ("Move fastboot over
  to ParseByteCount.") treat -S 0 and its absence the same:
  The target's max-download-size variable is queried.
- Older fastboot releases, e.g. 1:7.0.0+r33-1 shipped by Debian
  Stretch, instead completely disables sparse when -S 0, while not
  supplying it queries the target.

To have the fastbootdriver interoperate with both versions,
supply -S only, when there's a user set value, and otherwise
do not supply the option at all.

Fixes: #416, 1bf8952a ("driver: fastbootdriver: allow setting sparse option")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>