remote/coordinator: make resource object persistent
Prevously, we would create new ResourceEntry objects even for updates of
existing resources. To keep coordinator-specific information for
individual resources (such as whether they are acquired for a place),
the natural way is to keep the objects around.

Introduce a ResourceImport class analogous to ResourceExport (used by
the exporter), which also stores the resource path as exposed to the
clients. This also avoids recreating the resource path every time.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>