shelldriver: use remaining timeout after password
Instead of hard coding a 5 sec timeout, we can use whatever is left of
the global timeout for getting shell, when waiting for the shell after
inputting password.

This solves issues when the shell is delayed after inputting a password,
which can happen with slow devices or slow authentication systems.

Signed-off-by: Tobias Olausson <tol@hms.se>