usbvideodriver: add support for newer C920 Webcams
Newer Logitech C920 versions seem to have a different USB VID. They also
seem to support the newer UVC1.5 standard, which is not implemented by
Linux, therefore the h264 output is unfortunately not supported. This PR
adds a separate pipeline for these devices which use the MJPG output and
VA-API to convert to a h264 stream.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>