tests/conftest: read output from background processes
If the crossbar and coordinator produce a lot of output, they can fill
the kernel buffers connected to our pexpect spawn. In that case, they
block while writing and cause test failure.

Add a background thread which reads any output and pass it to stdout via
the pexpect logfile feature.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>