remote/client: add write-files command
Analogously to write-image, add a write-file command that can be used to
overwrite (or create anew) a single file in a file system located on a
USBStorageDriver. This is useful for faster iteration during bootloader
development when a bootloader is loaded from a file system.

Following usages are supported:

  labgrid-client write-files a b      # writes files a, b into /mnt/
  labgrid-client write-files -t a b c # writes files b, c into /mnt/a/
  labgrid-client write-files -T a b   # writes file  b      to /mnt/a

Where /mnt is some dynamic path on the, possibly remote, host exporting
the USBStorageDriver.

Signed-off-by: Stefan Wiehler <stefan.wiehler@missinglinkelectronics.com>
Co-authored-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Bastian Krause <bst@pengutronix.de>