util: expect: implement sendcontrol()
Sometimes it's useful to send Ctrl-C to interrupt a process or Ctrl-Z to
send a hung process to the background and kill it then (e.g.
gst-validate-1.0). This is often faster than transitioning to the
off/root state and back to the shell.

Although pexpect.spawn has a sendcontrol() method it relies on
ptyprocess which is not useful for labgrid's usecase. So implement it
here.

Signed-off-by: Bastian Krause <bst@pengutronix.de>