driver/sshdriver: add keepalive
Add a keepalive function to the SSHDriver, similar to the utils ssh
implementation. This is necessary since during tests the control master
may run into the configured 5 minute timeout, if multiple targets are
used and no commands are run on a target for some time. On the other
hand we don't want to keep the control master open longer after exit, so
we instead introduce a keepalive function which keeps the control master
running as long as the driver is not deactivated.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>