util/helper: add print_on_silent_log kwarg to ProcessWrapper.check_output()
If print_on_silent_log is True and the log level prevents process output
(info) logging ProcessWrapper's print_callback is registered before any
process output is processed and unregistered after the process
terminated.

This is meant to be set for interesting process output that is
understandable by the user without any context. Such an example is
dd's output used for 'labgrid-client write-image' (will be implemented
in a future commit).

Signed-off-by: Bastian Krause <bst@pengutronix.de>