util/helper: split process output on \r instead of \r\n
This allows processing output from commands that rewrite the same line
over and over again (e.g. 'dd status=progress').

The output of ProcessWrapper.check_output() does not change. Only
Callbacks are now called with output split on carriage returns ('\r'),
line breaks ('\n') are preserved. It is the responsibility of the
callback to handle this correctly.

The print_callback() now adds a leading \r again to allow printing simlar
output as is emitted by the original process. \n is preserved as is.

The logging_callback() makes sure the message is not empty after leading
and tailing line breaks are stripped.

Signed-off-by: Bastian Krause <bst@pengutronix.de>