sshdriver: ssh_prefix to list and adjust users
Convert ssh_prefix to a list and adjust all users. This was a string
previously, which led to get() and put() being broken on master, since
the string contained spaces. Convert it to a list, adjust the callers
and add new tests to test the SSHDriver against localhost. We also add a
new pytest option to supply the username, to make this work on travis.
Remove the test class used in the sshdriver tests while touching the file.

Fixes: 1d7b97b838f7 (driver/sshdriver: Allow whitespaces in filenames Prevent filenames with whitespaces from beeing split.)

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>