man/labgrid-exporter: remove useless sentence
This came up in the review of PR #583. First, the sentence is in the
section about configuration, but doesn't say anything about
configuration; second, it is not very helpful at all about how to
start ser2net; and third, the same statement is already made in
overview.rst, section "Exporter".

Signed-off-by: Roland Hieber <rhi@pengutronix.de>