labgrid/remote: remove explicit path from ser2net
Remove the explicit path from ser2net. Setting PATH correctly is for the
user to decide, and this allows overriding it externally with a locally
compiled version by modifying PATH. Instead try to resolve ser2net using
shutil.which and if not found fall back to /usr/bin/ser2net.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>