pytestplugin: add a --lg-initial-state= option to use during development
When repeatedly re-executing single test cases during development, the
time needed to bring the target in the proper state is usually much
larger than the actual test case duration. We can mitigate this by
allowing the user to override the initial state, and letting the
Strategy activate the matching drivers.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>