target: enable de/activate by string
self._reg_class_from_string() only retrieves the class, not the string.
Use the _get_driver call to retrieve the instance. Also add the name
parameter to enable usage in case multiple classes are bound to the same
target. While at it implement the same syntax for activate as well.

Also add tests for this feature.

Fixes: 103eb18014b4 ("target: rename lookup function")

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>