util/proxy: prefer isolated over LG_PROXY
If we want to connect to an isolated exporter, prefer a connection to
the exporter over the LG_PROXY variable. We expect the user to have a
properly setup SSH config to access the exporter, document this as a
note in the proxy documentation.

Fixes https://github.com/labgrid-project/labgrid/issues/620

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>