doc/usage: be a bit more verbose about what driver activation means
The concept of "activation" is mentioned a few times in the docs, but it
is explained nowhere. Try to change that by distilling the code from
labgrid.target.Target.activate() into an explanation with examples.
Thanks to TheMeaningfulEngineer for the car analogy.

Also be more concise about the difference between Resource objects and
the hardware that is represented by them.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>