sshdriver: include port in the ControlSocket name
Previously only the address was included in the ControlSocket name,
which leads to the SSHDriver killing the first connection by overwriting
the ControlSocket on the second connection. Include the port name in the
ControlSocket name to allow multiple connections to the same host on
different ports.

Fixes https://github.com/labgrid-project/labgrid/issues/658

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>