driver: qemudriver: prevent multiple -append QEMU arguments
Various configuration parameters of the QEmuDriver lead to QEMU getting
called with -append. This might not be obvious to the user. If the user
specifies -append in extra_args instead of using boot_args, the -append
argument in extra_args is ignored by QEMU. This happens because the last
occurrence of -append is used, all other occurrences are silently
ignored [1].

Add an explicit check to make sure no -append is given in extra_args, and
suggest the usage of boot_args which leads to only a single -append
being passed to QEMU.

[1] https://lists.gnu.org/archive/html/qemu-devel/2017-05/msg03214.html

Signed-off-by: Bastian Krause <bst@pengutronix.de>