target: always resolve strings via target_factory
The target_factory has a list of all imported drivers and resources whom
register themselves through the reg_driver or reg_resource decorator.
The target implemented an inferior version of this, which only tracked
bound drivers on the target. For all string to resource/driver resolve
cases, we always want to resolve against all registered
drivers/resources, not only those bound against the current target.
Replace and remove the _reg_class_from_string() method of the target
with class_from_string() method of the target_factory.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>