util/helper: multiple callback (un)registration
The pytester pytest plugin will setup multiple pytest sessions and calls
the pytest_configure multiple times. This leads to multiple registration
of the logging callback, which triggers the assert that the callback can
only be registered once. Remove the asserts and replace it with checks
to (un)register the callback only once.

Fixes https://github.com/labgrid-project/labgrid/issues/684

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>