usbstoragedriver: Use blockdev command to get disk size
Use the blockdev command to determine the size of a block device. This
command has the advantage that it directly calls ioctl(BLKGETSIZE64) on
the device, meaning it will work as long as the user has read permission
on the actual block device (e.g. /dev/sdb) instead of requiring
permissions to read the sysfs file.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>