target: duplicate bindings under differing names
The duplicate binding mechanism forbid the duplicate binding of the same
supplier:

  class DriverWithMultiA(Driver):
      bindings = {
        "res1": ResourceA,
        "res2": ResourceA,
      }

Is not allowed since the same resource is bound twice. We do howver have
Driver implementing two protocols, like the QEMUDriver:

  class QEMUDriver(ConsoleExpectMixin, Driver, PowerProtocol, ConsoleProtocol):

Here we do want to allow the binding of the QEMUDriver under differing
Names, namely binding it as a PowerProtocol and ConsoleProtocol.
Rework the code to keep a set of (requirement, supplier) to check
against and prevent binding against the same class with a duplicate
binding error, however allow it if the supplier was found under a
different requirement. This allows bindings like this:

  class BootStrategy(Strategy):
      bindings = {
        "power": "PowerProtocol",
        "console": "ConsoleProtocol"
      }

Where both are supplied by the same QEMUDriver.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>