docker: rename docker to dockerfiles
Since python3.3 python will not only import files with __init__.py, but
instead also imports directory trees which contain a single python file.
Therefore python running in the main repository directory will detect
docker as a module and imports it. This in turn shadows the real docker
module, if it is installed and breaks the tests docker import check.
Move the docker files into a new dockerfiles directory which does not
shadow the module name.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>