dockerfiles: Add yq and qemu-system to client container
Adds `yq` to the client container image. This tool is a front end for
`jq` that can parse yaml files, which is useful for writing CI pipelines
that are reactive to the contents of Labgrid environment files (e.g.
only reserve a place if one is specified in the environment).

Adds `qemu-system` to the client container, which allows testing using
qemu inside the container.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>