add flash scripting driver
Some SoC vendors have complex and SoC specific protocols and programs
that are required to flash software on their devices. Instead of trying
to implement drivers for all of these one-off implementations, implement
the surgical chain-saw approach and add support for sending a "flashing
script" to be run on the exporter which will flash the device.

A typical use case might be creating a self extracting flashing script
with e.g. makeself that contains the image files and all flashing
utilities.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>