add TFTP-/NFS-/HTTPProvider resources and drivers
This supports "staging" files to a remote TFTP/NFS/HTTP server for
access by the target. Files below the 'internal' path need to be visible
to the target below the corresponding 'external' path.

The actual server needs to be set up outside of labgrid for now. Running
and configuring the respective server could be done later in the
exporter without changes to the client.

The files are synchronized to the exporter via the ManagedFile helper
and then symlinked to the actual provider path.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>