shelldriver: Add post-login "settle" time
Some devices may have chatty consoles after a user logs in;
specifically, if a user logs into the root serial console of a target
using systemd that has not finished booting, boot messages will begin to
appear after login. These messages can interrupt prompt detection and
cause test timeouts. Add an API to wait for the device to "settle", e.g.
no output is seen on the console for a specified number of seconds and
add an option for the shelldriver to utilize it.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>