modbusdriver: retrieve exception information
We got a report of a user trying to access a modbus device with the
following error:

  Traceback (most recent call last):
  File "/opt/venvs/labgrid/lib/python3.9/site-packages/labgrid/remote/client.py", line 1638, in main
    args.func(session)
  File "/opt/venvs/labgrid/lib/python3.9/site-packages/labgrid/remote/client.py", line 814, in digital_io
    drv.set(False)
  File "/opt/venvs/labgrid/lib/python3.9/site-packages/labgrid/binding.py", line 100, in wrapper
    return func(self, *_args, **_kwargs)
  File "/opt/venvs/labgrid/lib/python3.9/site-packages/labgrid/driver/modbusdriver.py", line 36, in set
    raise ExecutionError('Could not write coil (code={})'.format(error_code))
  labgrid.driver.exception.ExecutionError: Could not write coil (code=7)

The error code is 7, indicating that an exception happened, however
labgrid did not retrieve the exception code, so no further debugging was
possible. Unify the modbusdriver error handling and retrieve the
exception if it happened. Also ignore the ACKNOWLEDGE and NONE exceptions.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>