driver/power/gude8031: Get rid of HTTP404 on set()
This commit changes the URL to an existing URL on the device.

It seems switching the ports on and off using the `cmd=1&p=N&s=M`
parameters works even if the URL does not exist.
But accessing an undefinded URL returns a HTTP404 that in turn causes an
exception.

Signed-off-by: Chris Fiege <c.fiege@pengutronix.de>