shelldriver: Move password handling into main poll loop
Moving the password handling into the main loop allows it to follow the
same timeout handling logic as other prompts. Specifically, if no output
is seen an empty newline will be sent instead of waiting indefinitely
for a timeout. This prevents test timeout errors when other output is
intermingled with the password prompting or the main prompt after
entering the password.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>