ubootdriver: Use a loop to access the u-boot console
Converts the u-boot driver to use a loop to try an login to the console,
just like the shelldriver does. This ensures that the driver waits for
the console prompt after interrupting and before attempting to enter a
command. Otherwise, if u-boot is not waiting at the prompt it may
occasionally miss the input, causing the test to fail.

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>