contrib/sync-places: Fix missing matches on tag on new places
The code was only adding matches and tags for places that were seen
while processing adding/removing places. This meant that brand-new
places did not get the matches and tags set. Fix this to set tags and
matches for all places specified in the config file, but only query the
existing places and tags for places that were seen (which prevents
dry-run from breaking)

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>