driver/power: Add backend driver for EG PMS2 LAN/WLAN
This driver adds support for the Energenie PMS2 LAN & WLAN programmable
power sockets. They are accessed over a web interface and require a
login before accessing the status of the socket or switching them on or off.
A password is required for the login in POST request and this driver
deliberately uses the default password, as the current state of labgrid
doesn't support password encryption, meaning the password would be
stored in plain text.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>
[r.czerwinski@pengutronix.de: add additional raise_for_status on responses]
Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>