driver: qemudriver: fix tool lookup for multiple calls to on_activate()
When the QEmuDriver is activated, then deactivated and activated again,
the tool lookup for "qemu_bin" is only done correctly for the first
activation. For all subsequent activations, the path is used incorrectly
as the tool label (key).

Fix that by storing the tool path in a different variable than the key,
so multiple activations work correctly.

Fixes: b3faf8e7 ("driver/qemudriver: configure paths and images via config")
Signed-off-by: Bastian Krause <bst@pengutronix.de>