strategy: wait for systemd operational state in shell states
The system might not be in operational state once the login prompt is
shown. Try to wait until the system is up and running for systemd-based
systems. Use run() instead of run_check() for non-systemd based systems
or if degraded/maintenance/stopping system states are reached.

This covers wider use-cases for the shipped strategies.

Signed-off-by: Bastian Krause <bst@pengutronix.de>