remote/client: add export subcommand
This adds support to export the low-level information used by the drivers to
external programs. To use this, you need to run labgrid-client with an
environment configuration (-c):

   $ labgrid-client -c foo.yaml export -
   Selected role main and place foo from configuration file
   export LG__CLIENT_PID=373996
   export LG__CONSOLE_HOST=rlaba-srv
   export LG__CONSOLE_PORT=52855
   export LG__CONSOLE_PROTOCOL=rfc2217
   export LG__CONSOLE_SPEED=115200
   export LG__ETH_HOST=raspberrypi
   export LG__ETH_IFNAME=''
   export LG__TFTP_EXTERNAL=foo/
   export LG__TFTP_HOST=raspberrypi
   export LG__TFTP_INTERNAL=/srv/tftp/foo/
   export LG__WIFI_HOST=raspberrypi
   export LG__WIFI_IFNAME=wlp1s0u1u3
   Waiting for CTRL+C or SIGTERM...

Writing to a file is supported as well:

   $ labgrid-client -c foo.yaml export test.env
   Selected role main and place foo from configuration file
   Exported to test.env
   Waiting for CTRL+C or SIGTERM...

When using this from a script, you will need to poll until the file is created
and the source the shell variables from it. The script can later stop
labgrid-client by sending using kill $LG__CLIENT_PID.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>