pytestplugin: fixtures: redirect stderr of git commands to /dev/null
labgrid's env fixture calls "git rev-parse HEAD" in each path configured
in the environments "paths" and "images" keys to record their
corresponding commit-ishes.

Because of that, errors like these are quite common:

  fatal: not a git repository (or any parent up to mount point /)
  Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).

This happens if the paths do not contain a git repository. Redirect
these errors to /dev/null instead of printing them without context,
confusing the users of the pytest plugin.

Signed-off-by: Bastian Krause <bst@pengutronix.de>